[Core] Fix opencensus protobuf bug (#31632)

Here's how the conflict happens.

We pull the opencensus-proto 0.3.0 for our bazel build.
opencensus-proto 0.3.0 generated python code (metric_pb2 and resource_pb2), which is used by reporter.proto.
We copy all python protobuf to python/ray/core/generated.
Now reporter_pb2 is available from python/ray/core/generated.
reporter_pb2 is not supposed to be imported from core workers (it is only for dashboard agent). However, this was imported (accidentally) from gcs_pubsub.py because pubsub.proto requires this module (which was unnecessary).

As a result, reporter_pb2 was imported to the core worker, which subsequently imports metric_pb2 and resource_pb2. When we import the opencensus proto using open telemetry, we import resource_pb2 again, which causes the conflict.

Theoretically, when the same package is imported (which is our case), protobuf is supposed to handle it. (I guess) This wasn't the case because our cpp version uses opencensus-proto 0.3.0, and the python only supports opencensus-proto 0.1.0 (omg https://pypi.org/project/opencensus-proto/#history).

There are 4 potential solutions.

[current PR] Make sure to not import reporter_pb2 from the core worker. It is the easiest fix and have the best compatibility story. Since we don't import reporter_pb2 (which is only for dashboard agent), there's no concern of conflict. Since we have unit tests now, we can avoid the future regressions.
Use downloaded opencensus-proto for all python generated code (instead of cpp one we built). This requires us to introduce opencensus-proto 0.1.0 as a dependency, so I avoided this approach.
Downgrade cpp opencensus-proto version to 0.1.0. This is not a very clean solution. What if opencensus releases 0.2.0?
Copy & paste opencensus-proto and build our own proto with different names. This can potentially have compatibility issue with opencensus library, but it is unlikely (cuz the project doesn't seem to be maintained). This is the fallback option if the current approach doesn't work.
