+++ python/ray/tests/test_memory_pressure.py
@@ -6,12 +6,16 @@ import time
 import pytest
 
 import ray
-from ray._private import test_utils
+from ray._private import (
+    ray_constants,
+)
+import ray._private.gcs_utils as gcs_utils
 from ray._private.test_utils import wait_for_condition, raw_metrics
 
 import numpy as np
 from ray._private.utils import get_system_memory
 from ray._private.utils import get_used_memory
+from ray.experimental.state.state_manager import StateDataSourceClient
 
 
 memory_usage_threshold = 0.65
@@ -20,6 +24,25 @@ memory_monitor_refresh_ms = 100
 expected_worker_eviction_message = (
     "Task was killed due to the node running low on memory"
 )
+
+
+def get_local_state_client():
+    hostname = ray.worker._global_node.gcs_address
+
+    gcs_channel = ray._private.utils.init_grpc_channel(
+        hostname, ray_constants.GLOBAL_GRPC_OPTIONS, asynchronous=True
+    )
+
+    gcs_aio_client = gcs_utils.GcsAioClient(address=hostname, nums_reconnect_retry=0)
+    client = StateDataSourceClient(gcs_channel, gcs_aio_client)
+    for node in ray.nodes():
+        node_id = node["NodeID"]
+        ip = node["NodeManagerAddress"]
+        port = int(node["NodeManagerPort"])
+        client.register_raylet_client(node_id, ip, port)
+        client.register_agent_client(node_id, ip, port)
+
+    return client
 
 
 @pytest.fixture
@@ -279,7 +302,7 @@ async def test_actor_oom_logs_error(ray_with_memory_monitor):
             oom_actor.allocate.remote(bytes_to_alloc, memory_monitor_refresh_ms * 3)
         )
 
-    state_api_client = test_utils.get_local_state_client()
+    state_api_client = get_local_state_client()
     result = await state_api_client.get_all_worker_info(timeout=5, limit=10)
     verified = False
     for worker in result.worker_table_data:
@@ -320,7 +343,7 @@ async def test_task_oom_logs_error(ray_with_memory_monitor):
             )
         )
 
-    state_api_client = test_utils.get_local_state_client()
+    state_api_client = get_local_state_client()
     result = await state_api_client.get_all_worker_info(timeout=5, limit=10)
     verified = False
     for worker in result.worker_table_data:
